Incoming oneOf collection is now passed directly as argument.
Related tickets #208
