+++ src/test/java/am/ik/yavi/Message.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (C) 2018-2022 Toshiaki Maki <makingx@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package am.ik.yavi;
+
+public class Message {
+	private final String text;
+	private final Color color;
+
+	public Message(String text, Color color) {
+		this.text = text;
+		this.color = color;
+	}
+
+	public String getText() {
+		return text;
+	}
+
+	public Color getColor() {
+		return color;
+	}
+}
